requirements.txt
gql==3.4.0
python_dateutil==2.8.2
requests==2.22.0
web3==5.30.0
python-dotenv==1.0.0